requirements.txt
flask
flask_sqlalchemy
requests
discord.py
gunicorn